refactored resize
